feat(new-site): Initial Commit from Template
Next.js offers a lot of promise, so I've been exploring variations of
static site templates with it. For my personal site, I know I want to be
able to use a mostly markdown but also a certain amount of custom
components.

This initial commit includes a template that I've copied and includes:

- Swaps preact instead of react to slim down initial bundle size
- Includes mdx rendering, but files can also import components as
  necessary to display more advanced elements
- Include the `Link` in place of anchors to make page linking faster
- Very basic layouts right now, just enough to display a page and an
  introductory emoji. (lots to build from)
- Fathom analytics: I'd like to keep some basic analtyics about
  pageviews (and what attracts more visitors, etc): fathom is wired into
  the root App