feat(day10): Day10 post, whois, a placeholder
In the next couple of days I hope to get a placeholder page online that
can share some of the content I'm building for the CollabClass.

But in the meantime, for tonight I'm just adding a day10 post and a
couple of misc typo updates.

Also updated the core next.config.js to only use core mdx module, not
the mdx-enhanced module