feat(content): Update content up to day 13
Catching up on my post for the day before, plus an edit from the day
before that. And because I want to deploy this today I need the results
of todays project as well.

I think today I'm going to keep it simple and focus on exploring what it
looks like to load files via glob patterns (for bulk importing mdx
files).