feat(list-pages): Experiments with list pages
Next with mdx gives us a lot of options for how we present the content
to our users. However, one issue that is still unclear is how to go
about rendering LISTS of mdx content. Babel glob imports are an easy
solution, but we still need to be careful about how we pass that data
into individual pages.